[#174952489] Add base error class and response errors
