+++ lib/dhl_uk.rb
@@ -3,7 +3,9 @@
 require 'json'
 require 'faraday'
 
+require 'dhl_uk/errors/error'
 require 'dhl_uk/errors/auth_token_error'
+require 'dhl_uk/errors/response_error'
 
 require 'dhl_uk/version'
 require 'dhl_uk/config'
